updated sitemap-generator
